requirements.txt
aiomysql==0.2.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
bcrypt==3.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
databases==0.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.110.1
h11==0.14.0
idna==3.7
Mako==1.3.5
MarkupSafe==2.1.5
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.1
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.29.0
